Updated the bold text for event
